+++ app/models/api_test_case.py
@@ -11,7 +11,7 @@
 """
 from typing import List
 from uuid import uuid4
-from sqlalchemy import Column, String, INT, TEXT, SMALLINT, UniqueConstraint
+from sqlalchemy import Column, String, SMALLINT, TEXT, SMALLINT, UniqueConstraint
 from app.enums.ByteSizeEnum import ByteSizeEnum
 from app.enums.SysVarEnum import PikaGlobalVarEnum
 from app.models.api_testcase_out_parameters import ApiTestCaseOutParametersModel
@@ -26,7 +26,7 @@ class ApiTestCaseModel(LargeBaseModel):
         'directory_id', 'name'), {"comment": "测试用例表"})
     name = Column(String(ByteSizeEnum.LENGTH_50), index=True, comment="名称")
     request_type = Column(
-        INT, default=1, comment="请求类型 1: http 2: grpc 3: dubbo")
+        SMALLINT, default=1, comment="请求类型 1: http 2: grpc 3: dubbo")
     url = Column(TEXT, nullable=False, comment="请求url")
     request_method = Column(String(ByteSizeEnum.LENGTH_12), nullable=True,
                             comment="请求方式, 如果非http可为空")
@@ -34,11 +34,11 @@ class ApiTestCaseModel(LargeBaseModel):
     base_path = Column(String(ByteSizeEnum.LENGTH_24), comment="请求base_path")
     body = Column(TEXT, comment="请求body")
     body_type = Column(
-        INT, comment="请求类型, 0: none 1: json 2: form 3: x-form 4: binary 5: GraphQL")
+        SMALLINT, comment="请求类型, 0: none 1: json 2: form 3: x-form 4: binary 5: GraphQL")
     directory_id = Column(BinaryUUID,
                           default=uuid4, comment="所属目录")
     tag = Column(String(ByteSizeEnum.LENGTH_64), comment="用例标签")
-    status = Column(INT, comment="用例状态: 1: 调试中 2: 暂时关闭 3: 正常运作")
+    status = Column(SMALLINT, comment="用例状态: 1: 调试中 2: 暂时关闭 3: 正常运作")
     priority = Column(String(ByteSizeEnum.LENGTH_03), comment="用例优先级: p0-p3")
     case_type = Column(SMALLINT, comment="0: 普通用例 1: 前置用例 2: 数据工厂")
     out_parameters: List[ApiTestCaseOutParametersModel] = None
