+++ app/models/role.py
@@ -10,7 +10,7 @@
 @Desc    :  角色配置表
 """
 from uuid import uuid4
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, SMALLINT, String
 from app.core.handler.sqlbin_uuid import BinaryUUID
 from app.enums.ByteSizeEnum import ByteSizeEnum
 from app.enums.SysVarEnum import PikaGlobalVarEnum
@@ -24,11 +24,11 @@ class RoleModel(TimestampBaseModel):
                 default=uuid4, nullable=False, primary_key=True)
     name = Column(String(ByteSizeEnum.LENGTH_64),
                   nullable=True, comment='菜单名称', index=True)
-    role_type = Column(Integer, server_default='10', nullable=False,
+    role_type = Column(SMALLINT, server_default='10', nullable=False,
                        comment='权限类型,10菜单权限,20用户组权限', index=True)
     menus = Column(String(ByteSizeEnum.LENGTH_255),
                    nullable=True, comment='菜单列表', index=True)
-    status = Column(Integer, server_default='10',
+    status = Column(SMALLINT, server_default='10',
                     nullable=True, comment='状态 10 启用 20 禁用')
     description = Column(String(ByteSizeEnum.LENGTH_600),
                          default=None, comment="备注信息")
